feat(storybook): add CorrectScore.stories.js
